Allow MustacheReplacer#replace to take a block
Enhance MustacheReplacer#replace to optionally accept a block. This
block, when passed, yields the current value being replaced as well as
the path to that value.

The last expression returned from the block will be the new path in the
manifest.

This allows a caller of MustacheReplacer to hook in and do something
with the value and path before the replace happens.