Enhance SanitizeExecutor#execute to take a block
SanitizeExecutor#execute optionally accepts a block. This block, when
given, yields access to the MustacheReplacer and its block. That is, the
caller has access to the current value being replaced as well as the
path to that value.

This allows someone calling the SanitizeExecutor programatically to hook
in and do something with the paths and values as part of the replacing
process.